Refactor(추천 시스템) : 이벤트 핸들러가 서비스 를 의존하도록 수정
